postgres-db-agent/requirements.txt
mcp==1.19.0
python-dotenv==1.1.1
google-adk==1.17.0
a2a-sdk==0.3.10
click==8.2.1
strands-agents==1.13.0
strands-agents-tools==0.0.1
strands-agents[openai]==1.13.0
strands-agents[litellm]==1.13.0
asyncpg==0.30.0